// packages/codegen/src/output/database/src/mongoose/interfaces/iUserAgentCreateInput.ts
// THIS CODE WAS AUTOMATICALLY GENERATED
import {databaseTypes} from 'types';
import {Types as mongooseTypes} from 'mongoose';

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface IUserAgentCreateInput
  extends Omit<databaseTypes.IUserAgent, '_id' | 'createdAt' | 'updatedAt'> {}